+++ styles/assignmentsTable.module.css
@@ -1,0 +1,58 @@
+.table {
+    border-collapse: collapse;
+    width: 100%;
+    font-size: large;
+    table-layout: auto;
+}
+.table td,
+.table th {
+    border: 2px solid var(--text);
+    padding: 4px;
+    text-align: center;
+}
+.headerRow th div {
+    display: flex;
+    align-items: center;
+    align-content: center;
+    justify-content: center;
+}
+.headerRow h4 {
+    margin: 0;
+    user-select: none;
+}
+.sortedAsc .ascArrow {
+    display: inline;
+}
+.sortedDesc .dscArrow {
+    display: inline;
+}
+.ascArrow,
+.dscArrow {
+    display: none;
+}
+.linkCell {
+    max-width: 10vw;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+}
+.buttonCell {
+    position: relative;
+    height: 0;
+    white-space: nowrap;
+    width: 0;
+}
+.tableButton {
+    position: relative;
+    top: 50%;
+    transform: translateY(-50%);
+    padding: 0 0.5ch;
+}
+@media (pointer: fine) {
+    .headerRow th:hover {
+        cursor: pointer;
+    }
+    .tableButton:hover {
+        cursor: pointer;
+    }
+}
