+++ styles/globals.css
@@ -3,7 +3,6 @@
     --text: #272729;
     --font: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen, Ubuntu,
         Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
-    --overdue: #ff0000;
 }
 
 body {
